It worked! It just took forever to compile the files it needed.
I'm not sure if there errors that were thrown at the beginning will
become a problem, later, but I was able to load the `(web client)`
module so I think it should run well enough to run Artanis.

It probably built fine, before, but it's better to build things than
try to install them through `apt` so we'll keep the last commit.

I think the default load path is under `/usr` so let's set that so,
hopefully, things work correctly when Guile bindings for GnuTLS get
built.

And, then, it'll be set for the user when they start the dyno.