So, naturally, Heroku buildpacks have a timeout set; frankly I'm surprised Guile made the cut (of course, that isn't including all the compiling that it does on first run…).
So we'll do it this way: this'll be a Guile buildpack. We can nix the
autogen since that's for GnuTLS (and we can try building that from
source, in _that_ buildpack). And then we'll have a third buildpack to
build Artanis.

I'm gonna keep the other packages commented until I can transfer them
to their own repos.