Alright; so that quasi-worked: it got some but not everything.
So let's go nuclear and target every file and the specific string we
want to replace.

We'll include up until the `/.apt/usr` part because we know that was
in the original prefix so we should be safe to still target all the
ones we want and because that'll ensure we don't accidentally hit a
String we're not looking for (not that that seems particularly
likely…; if we still are missing some instances, we'll chop the
`/.apt/usr` and try again).