Fuck it; this is too much to keep track of.
If I can do so for each, I'd like to but I can check that, later;
I'll, at least, just have the Artanis buildpack save the entire `.apt`
directory so things are just copied over rather than trying to
reinstall already built sources.

I guess the previous buildpacks just won't do anything, if they see
the directory in the cache; it's not clean but it works (which is
pretty much how this whole thing's been, really).