This is what we must reside ourselves to.
So, since other attempts to cache didn't work for each individual
buildpack, we'll just cache things for the sources at the end, with
the Artanis buildpack.

So, if we find that the `STACK`'s changed and Artanis doesn't seem to
be installed, just build from scratch. However, – if the stack didn't
change but Artanis isn't there – an APT package is an APT package and
there's no reason to reinstall them if the `STACK`'s still the same as
last time.

For the sources, – if Artanis isn't installed – build 'em; if it is,
don't do anything: the Artanis buildpack'll restore them from the cache.